prefix all paths with /data
