+++ app/app.py
@@ -7,6 +7,7 @@ from sqlite3 import Error
 from flask import Flask
 from constants import DATASET_PATH, DEFAULT_DATABASE
 from databasemanager import DatabaseManager
+from routes.data_route import data_blueprint
 
 
 def create_app(test_config=None):
@@ -25,29 +26,6 @@ def create_app(test_config=None):
     path = Path(__file__).parent / DATASET_PATH
     db_manager.import_dataset(path)
 
-    @app.route("/<string:name>/average", methods=["GET"])
-    def get_average_undernourishment_by_name(name):
-        """Return the average undernourishment for the given country"""
-        try:
-            data = db_manager.get_data_by_name(name.lower())
-        except Error as error:
-            return {
-                'error': {
-                    f'Error fetching data: {error}'
-                }
-            }, 500
-
-        undernourishments = [row_data.undernourishment for row_data in data]
-
-        if undernourishments != []:
-            return {
-                'data': {
-                    'average': sum(undernourishments) / len(undernourishments)
-                }
-            }, 200
-
-        return {
-            'error': 'No data found for name: ' + name
-        }, 404
+    app.register_blueprint(data_blueprint, url_prefix='/data')
 
     return app
