+++ tests/test_average_undernourishment.py
@@ -27,12 +27,12 @@ class TestAverageUndernourishment:
 
     def test_get_average_undernourishment_by_name(self, client):
         """Tests get_average_undernourishment_by_name() to return correct average."""
-        res = client.get('/World/average')
+        res = client.get('/data/World/average')
         assert 10.168420992399517 == res.get_json()['data']['average']
         assert res.status_code == 200
 
     def test_get_average_undernourishment_by_name_none(self, client):
         """Tests get_average_undernourishment_by_name() to return error when name
         doesn't exist."""
-        res = client.get('/Lalaland/average')
+        res = client.get('/data/Lalaland/average')
         assert res.status_code == 404
